move collection utils to localstack.utils.collections
