+++ localstack/utils/generic/file_utils.py
@@ -1,4 +1,2 @@
-from localstack.utils.files import cache_dir, parse_config_file
-
 # DEPRECATED: use localstack.utils.files
-__all__ = ["parse_config_file", "cache_dir"]
+from ..files import cache_dir, parse_config_file  # noqa
